feat(node): making a interface for node class
