+++ BinaryTree/src/binary_node.h
@@ -1,0 +1,9 @@
+#include "interface_node.h"
+
+class BinaryNode final : public INode
+{
+    public:
+        int data;
+        BinaryNode *left;
+        BinaryNode *right;
+};
